feat: 유효성검사 및 예외처리

+++ src/main/java/gift/Controller/ProductViewController.java
@@ -3,6 +3,8 @@ package gift.Controller;
 import gift.DTO.Product;
 import gift.DAO.ProductRepository;
 import org.springframework.stereotype.Controller;
+import jakarta.validation.Valid;
+import org.springframework.validation.BindingResult;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,7 +30,11 @@ public class ProductViewController {
     }
 
     @PostMapping("/add")
-    public String addProduct(@ModelAttribute Product product) {
+    public String addProduct(@Valid @ModelAttribute Product product, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return "product-form";
+        }
+
         productRepository.save(product);
         return "redirect:/products";
     }
@@ -41,11 +47,14 @@ public class ProductViewController {
     }
 
     @PostMapping("/edit/{id}")
-    public String editProduct(@PathVariable("id") Long id, @ModelAttribute Product updatedProduct) {
+    public String editProduct(@PathVariable("id") Long id, @Valid @ModelAttribute Product updatedProduct, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return "product-form";
+        }
         // 기존 제품을 조회
         Product product = productRepository.findById(id);
         if (!product.getId().equals(updatedProduct.getId())) {
-            // ID가 변경된 경우 기존 제품을 삭제하고 새 제품을 추가
+            // ID변경이 잘 안 돼서 추가한 코드.
             productRepository.deleteById(id);
         }
